requirements.txt
discord.py
urlopen
Request
bs4
urllib
random